Address review comments on path validation and OS platform caching
